tests: refactored tests that now all pass. Closes #9. Related to PR #1178
