+++ tests/test_visitors/test_tokenize/test_primitives/test_string_tokens/test_string_modifier.py
@@ -117,46 +117,3 @@ def test_uppercase_prefix_multiple(
         UppercaseStringModifierViolation,
         RawStringNotNeededViolation,
     ])
-
-
-@pytest.mark.parametrize('raw_strings', [
-    "{0}'Carriage return \r'",
-    "{0}'Escaped \' quote'",
-    "{0}'Escaped ab \t'",
-    "{0}'Escaped backslash \\'",
-])
-def test_necessary_raw_string(
-    parse_tokens,
-    assert_errors,
-    default_options,
-    raw_strings,
-):
-    """Ensures that correct usage of raw string works."""
-    string = raw_strings.format('r')
-    file_tokens = parse_tokens(string)
-
-    visitor = WrongStringTokenVisitor(default_options, file_tokens=file_tokens)
-    visitor.run()
-
-    assert_errors(visitor, [])
-
-
-@pytest.mark.parametrize('raw_strings', [
-    "{0}'No escaped caracter'",
-    '{0}"Here neither"',
-    "{0}'''Not here as well'''",
-])
-def test_unnecessary_raw_string(
-    parse_tokens,
-    assert_errors,
-    default_options,
-    raw_strings,
-):
-    """Ensures that usage of raw string is forbidden if no backslash."""
-    string = raw_strings.format('r')
-    file_tokens = parse_tokens(string)
-
-    visitor = WrongStringTokenVisitor(default_options, file_tokens=file_tokens)
-    visitor.run()
-
-    assert_errors(visitor, [RawStringNotNeededViolation])
